more migrating of refactoring
extracting cahnges from github mess...
