+++ GEECS-PythonAPI/geecs_python_api/controls/data_acquisition/data_acquisition.py
@@ -24,6 +24,8 @@ import geecs_python_api.controls.interface.message_handling as mh
 
 from image_analysis.utils import get_imaq_timestamp_from_png
 
+from .utils import get_full_config_path  # Import the utility function
+
 
 class Sounds:
     def __init__(self):
@@ -62,142 +64,14 @@ class Sounds:
 # Example usage of the Sounds class
 sounds = Sounds()
 
-
-class ConfigManager:
-    def __init__(self):
-        """
-        Initialize the ConfigManager with the base directory for configurations.
-        The base directory is set to be relative to the directory where this script/module resides.
-        """
-        # Get the path of the current file (where this class is defined)
-        current_dir = Path(__file__).parent
-
-        # Set the base directory to be the 'configs' directory relative to the current directory
-        self.base_dir = current_dir / 'configs'
-
-        # Ensure base_dir exists
-        if not self.base_dir.exists():
-            raise FileNotFoundError(f"The base config directory {self.base_dir} does not exist.")
-
-        self.experiment_dir = None
-
-    def set_experiment_dir(self, experiment: str):
-        """
-        Set the experiment directory based on the subdirectory name.
-
-        Args:
-            experiment (str): Name of the subdirectory for the experiment (e.g., 'HTU').
-        """
-        self.experiment_dir = self.base_dir / experiment
-        if not self.experiment_dir.exists():
-            raise FileNotFoundError(f"The experiment directory {self.experiment_dir} does not exist.")
-
-    def get_config_path(self, config_file: str) -> Path:
-        """
-        Get the full path of the specified configuration file in the current experiment directory.
-
-        Args:
-            config_file (str): The configuration file to retrieve.
-
-        Returns:
-            Path: The full path to the config file.
-        """
-        if not self.experiment_dir:
-            raise ValueError("Experiment directory is not set. Call set_experiment_dir() first.")
-
-        config_path = self.experiment_dir / config_file
-        if not config_path.exists():
-            raise FileNotFoundError(f"The config file {config_path} does not exist.")
-
-        return config_path
-
-    def visa_config_generator(self, visa_key, diagnostic_type):
-
-        # input_filename = '../../geecs_python_api/controls/data_acquisition/configs/HTU/visa_plunger_lookup.yaml'
-
-        input_filename = self.get_config_path('visa_plunger_lookup.yaml')
-
-        with open(input_filename, 'r') as file:
-            visa_lookup = yaml.safe_load(file)
-
-        device_info = visa_lookup[visa_key]
-
-        # Define the VisaEBeam camera dynamically based on the visa_key
-        visa_ebeam_camera = f"UC_VisaEBeam{visa_key[-1]}"  # Extracts the last number from visa_key (e.g., visa1 -> UC_VisaEBEam1)
-
-        if diagnostic_type == 'energy':
-            description = f"collecting data on {visa_key}EBeam and U_FELEnergyMeter"
-            setup_steps = [
-                {'action': 'execute', 'action_name': 'remove_visa_plungers'},
-                {'device': device_info['device'], 'variable': device_info['variable'], 'action': 'set', 'value': 'on'},
-                {'device': 'U_Velmex', 'variable': 'Position', 'action': 'set',
-                 'value': device_info['energy_meter_position']}
-            ]
-            devices = {
-                visa_ebeam_camera: {
-                    'variable_list': ["timestamp"],
-                    'synchronous': True,
-                    'save_nonscalar_data': True
-                },
-                'U_FELEnergyMeter': {
-                    'variable_list': ["Python Results.ChA", "timestamp"],
-                    'synchronous': True,
-                    'save_nonscalar_data': True
-                }
-            }
-
-        elif diagnostic_type == 'spectrometer':
-            description = f"collecting data on {visa_key}EBeam and U_Spectrometer"
-            setup_steps = [
-                {'action': 'execute', 'action_name': 'remove_visa_plungers'},
-                {'device': device_info['device'], 'variable': device_info['variable'], 'action': 'set', 'value': 'on'},
-                {'device': 'U_Velmex', 'variable': 'Position', 'action': 'set',
-                 'value': device_info['spectrometer_position']}
-            ]
-            devices = {
-                visa_ebeam_camera: {
-                    'variable_list': ["timestamp"],
-                    'synchronous': True,
-                    'save_nonscalar_data': True
-                },
-                'UC_UndulatorRad2': {
-                    'variable_list': ["MeanCounts", "timestamp"],
-                    'synchronous': True,
-                    'save_nonscalar_data': True
-                }
-            }
-
-        # Constructing the YAML structure
-        output_data = {
-            'Devices': devices,
-            'scan_info': {
-                'description': description
-            },
-            'setup_action': {
-                'steps': setup_steps
-            }
-        }
-
-        # Writing to a YAML file
-
-        output_filename = input_filename.parent / f'{visa_key}_{diagnostic_type}_setup.yaml'
-        with open(output_filename, 'w') as outfile:
-            yaml.dump(output_data, outfile, default_flow_style=False)
-
-        # print(f"YAML file {output_filename} generated successfully!")
-        return output_filename
-
-
 class ActionManager:
     def __init__(self, experiment_dir: str):
         # Initialize the ConfigManager within ActionManager
         self.config_manager = ConfigManager()  # Automatically initialized
         if experiment_dir is not None:
-            self.config_manager.set_experiment_dir(experiment_dir)  # Set the experiment directory
-
-            # Store the experiment directory path and the full path to actions.yaml
-            self.experiment_dir = self.config_manager.experiment_dir
-            self.actions_file_path = self.experiment_dir / 'actions.yaml'  # Path to actions.yaml
+            
+            # Use the utility function to get the path to the actions.yaml file
+            self.actions_file_path = get_full_config_path(experiment_dir, 'actions.yaml')
 
             # Dictionary to store instantiated GeecsDevices
             self.instantiated_devices = {}
@@ -301,25 +175,17 @@ class DeviceManager:
         self.event_driven_observables = []  # Store event-driven observables
         self.async_observables = []  # Store asynchronous observables
         self.non_scalar_saving_devices = []  # Store devices that need to save non-scalar data
-        self.base_config_file_path = []
-        self.composite_variables = []
-        self.config_manager = ConfigManager()  # Automatically initialized
 
         if experiment_dir is not None:
-            # Initialize the ConfigManager within DeviceManager
-            self.config_manager.set_experiment_dir(experiment_dir)  # Set the experiment directory
+            # Set the experiment directory
+            self.experiment_dir = experiment_dir
 
-            # Load paths for required config files using the ConfigManager
-            self.base_config_file_path = self.config_manager.get_config_path('base_monitoring_devs.yaml')
-            self.composite_variables_file_path = self.config_manager.get_config_path('composite_variables.yaml')
-
+            # Use self.experiment_dir when calling the utility function
+            self.base_config_file_path = get_full_config_path(self.experiment_dir, 'base_monitoring_devs.yaml')
+            self.composite_variables_file_path = get_full_config_path(self.experiment_dir, 'composite_variables.yaml')
+           
             # Load composite variables from the file
             self.composite_variables = self.load_composite_variables(self.composite_variables_file_path)
-
-        # Placeholder for scan description
-        self.scan_base_description = None
-        self.scan_parameters = None
-        self.scan_setup_action = None
 
     def load_composite_variables(self, composite_file):
         """
@@ -363,7 +229,7 @@ class DeviceManager:
         self.load_base_config()
 
         # Load the specific config for the experiment
-        config_path = self.config_manager.get_config_path(config_filename)
+        config_path = get_full_config_path(self.experiment_dir, config_filename)
         with open(config_path, 'r') as file:
             config = yaml.safe_load(file)
         logging.info(f"Loaded configuration from {config_path}")
@@ -418,10 +284,6 @@ class DeviceManager:
 
         logging.info(f"Devices loaded: {self.devices.keys()}")
 
-    def load_composite_variables(self, composite_file):
-        with open(composite_file, 'r') as file:
-            return yaml.safe_load(file).get('composite_variables', {})
-
     def is_statistic_noscan(self, variable_name):
         return variable_name in ('noscan', 'statistics')
 
@@ -445,6 +307,7 @@ class DeviceManager:
         # Iterate over each component and evaluate its relation
         for comp in components:
             relation = comp['relation'].replace("composite_var", str(value))  # Replace the placeholder
+            logging.info(f"Evaluating relation: {relation}")
             evaluated_value = eval(relation)  # Evaluate the relationship to get the actual value
             variables[f"{comp['device']}:{comp['variable']}"] = evaluated_value
